requirements.txt
Django
djangorestframework
psycopg2-binary
aiomultiprocess
factory-boy

